test: replace compatibility check script with an e2e test
Compatibility script was using the library sources,
and because of that depended on the @aws-sdk/types version installed on the root project.
That was in conflict with the AWS SDK client versions installed by the compatibility script.

E2E tests deploy package to a local npm registry (Verdaccio),
install the package from there and run unit tests with it.