feat!: infer Client type in AwsStub class
The AwsStub class was not intended to be used as mockClient() result type
as it required multiple generic parameters from the AWS SDK Client class.
Instead, AwsClientStub type taking only the Client type as generic was intended to be used.

However, TypeScript showed return type of mockClient() as AwsStub,
which lead users to use it anyway.

This change removes AwsClientStub type and makes the SDK Client type
the only required generic parameter of AwsStub class, making it easy to use.